Use fastcall calling convention to make debug builds work
Changed calling conventions of several functions from stdcall
to fastcall so we don't need to use inline asm to get to the
this pointer in ecx. The inline asm didn't work in debug mode
anyway, because stack checking code in the function prologue
clobbered ecx before the inline asm was executed.

This article explains how to use fastcall when hooking thiscall functions:
https://tresp4sser.wordpress.com/2012/10/06/how-to-hook-thiscall-functions/