Temp commit that seems to work
The problem with this version is that it inserts the fragment
in the level8 function while the packet is being built up.
This means that the caller of the level8 function is unaware
of this change. Consequently when the terminator bit is added
it is probably added in the wrong place.

A better plan is to only prepare things while the packet is
being built up and do the actual insertion from the sendto
function. I'm gonna try that next.